requirements.txt
matplotlib==3.3.4
numpy==1.23.3
pandas==1.5.0
Pillow==11.0.0
scikit_learn==1.5.2
tensorflow==2.10.0
opencv-python==4.5.5.64
llvmlite==0.43.0
numba==0.60.0